Move SourceLocation to evmcore
