+++ test/Assembly.cpp
@@ -24,7 +24,7 @@
 #include <iostream>
 #include <boost/test/unit_test.hpp>
 #include <libdevcore/Log.h>
-#include <libdevcore/SourceLocation.h>
+#include <libevmcore/SourceLocation.h>
 #include <libsolidity/Scanner.h>
 #include <libsolidity/Parser.h>
 #include <libsolidity/NameAndTypeResolver.h>
